software/src/requirements.txt
polars-lts-cpu==1.27.1
polars-ds==0.3.5
polars-hash==0.5.4
msgspec==0.18.6
PyYAML==6.0.2
